# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Order Manager API",
    "description": "API per la gestione degli ordini",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": ["http", "https"],
  "paths": {
...
    "/sales": {
      "get": {
        "tags": ["Sales"],
        "summary": "Ottiene la lista delle vendite",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "type": "string",
            "description": "Filtra per stato (pending, confirmed, shipped, delivered, cancelled)"
          },
          {
            "name": "customer_email",
            "in": "query",
            "type": "string",
            "description": "Filtra per email del cliente"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista delle vendite recuperata con successo",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Sale"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Sales"],
        "summary": "Crea una nuova vendita",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "customer_name": {
                  "type": "string",
                  "description": "Nome del cliente"
                },
                "customer_email": {
                  "type": "string",
                  "format": "email"
                },
                "customer_address": {
                  "type": "string"
                },
                "customer_phone": {
                  "type": "string"
                },
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "item_id": {
                        "type": "integer",
                        "description": "ID del prodotto"
                      },
                      "quantity": {
                        "type": "integer",
                        "description": "Quantità da vendere"
                      },
                      "unit_price": {
                        "type": "number",
                        "format": "float",
                        "description": "Prezzo unitario di vendita"
                      }
                    },
                    "required": ["item_id", "quantity", "unit_price"]
                  }
                },
                "notes": {
                  "type": "string",
                  "description": "Note aggiuntive"
                },
                "status": {
                  "type": "string",
                  "description": "Stato iniziale della vendita",
                  "default": "pending",
                  "enum": [
                    "pending",
                    "confirmed",
                    "shipped",
                    "delivered",
                    "cancelled"
                  ]
                }
              },
              "required": ["customer_name", "items"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Vendita creata con successo",
            "schema": {
              "$ref": "#/definitions/Sale"
            }
          },
          "400": {
            "description": "Dati mancanti o non validi"
          },
          "404": {
            "description": "Prodotto non trovato"
          }
        }
      }
    },
    "/sales/{sale_id}": {
      "get": {
        "tags": ["Sales"],
        "summary": "Ottiene i dettagli di una vendita",
        "parameters": [
          {
            "name": "sale_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Dettagli vendita recuperati con successo",
            "schema": {
              "$ref": "#/definitions/Sale"
            }
          },
          "404": {
            "description": "Vendita non trovata"
          }
        }
      },
      "put": {
        "tags": ["Sales"],
        "summary": "Aggiorna una vendita",
        "parameters": [
          {
            "name": "sale_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "enum": [
                    "pending",
                    "confirmed",
                    "shipped",
                    "delivered",
                    "cancelled"
                  ]
                },
                "customer_name": {
                  "type": "string"
                },
                "customer_email": {
                  "type": "string",
                  "format": "email"
                },
                "customer_address": {
                  "type": "string"
                },
                "customer_phone": {
                  "type": "string"
                },
                "notes": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendita aggiornata con successo",
            "schema": {
              "$ref": "#/definitions/Sale"
            }
          },
          "400": {
            "description": "Dati non validi o stato non valido"
          },
          "404": {
            "description": "Vendita non trovata"
          }
        }
      },
      "delete": {
        "tags": ["Sales"],
        "summary": "Elimina una vendita",
        "parameters": [
          {
            "name": "sale_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Vendita eliminata con successo"
          },
          "400": {
            "description": "Non è possibile eliminare una vendita non in stato pending"
          },
          "404": {
            "description": "Vendita non trovata"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Sale": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "customer_name": {
          "type": "string"
        },
        "customer_email": {
          "type": "string",
          "format": "email"
        },
        "customer_address": {
          "type": "string"
        },
        "customer_phone": {
          "type": "string"
        },
        "date": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "enum": ["pending", "confirmed", "shipped", "delivered", "cancelled"]
        },
        "total_amount": {
          "type": "number",
          "format": "float"
        },
        "notes": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SaleItem"
          }
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "SaleItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "item_id": {
          "type": "integer"
        },
        "item_name": {
          "type": "string"
        },
        "item_sku": {
          "type": "string"
        },
        "quantity": {
          "type": "integer"
        },
        "unit_price": {
          "type": "number",
          "format": "float"
        },
        "total_price": {
          "type": "number",
          "format": "float"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}